requirements.txt
numpy==1.24.3
cupy-cuda12x==12.1.0